+++ src/material/texture.rs
@@ -6,7 +6,7 @@ use std::fmt::Debug;
 use std::sync::Arc;
 
 pub trait Sample {
-    fn sample(&self, u: f32, v: f32, p: Vector3<f32>) -> Vector3<f32>;
+    fn sample(&self, u: f64, v: f64, p: Vector3<f64>) -> Vector3<f64>;
 }
 
 #[derive(Debug, Clone)]
@@ -18,7 +18,7 @@ pub enum Texture {
 }
 
 impl Sample for Texture {
-    fn sample(&self, u: f32, v: f32, p: Vector3<f32>) -> Vector3<f32> {
+    fn sample(&self, u: f64, v: f64, p: Vector3<f64>) -> Vector3<f64> {
         match self {
             Texture::SolidColor(texture) => texture.sample(u, v, p),
             Texture::Checkered(texture) => texture.sample(u, v, p),
@@ -30,37 +30,37 @@ impl Sample for Texture {
 
 #[derive(Debug, Clone)]
 pub struct SolidColor {
-    pub albedo: Vector3<f32>,
+    pub albedo: Vector3<f64>,
 }
 
 impl SolidColor {
-    pub fn texture(albedo: Vector3<f32>) -> Texture {
+    pub fn texture(albedo: Vector3<f64>) -> Texture {
         Texture::SolidColor(SolidColor { albedo })
     }
 }
 
 impl Sample for SolidColor {
-    fn sample(&self, _: f32, _: f32, _: Vector3<f32>) -> Vector3<f32> {
+    fn sample(&self, _: f64, _: f64, _: Vector3<f64>) -> Vector3<f64> {
         self.albedo
     }
 }
 
 #[derive(Debug, Clone)]
 pub struct Checkered {
-    pub even: Vector3<f32>,
-    pub odd: Vector3<f32>,
-    pub scale: f32,
+    pub even: Vector3<f64>,
+    pub odd: Vector3<f64>,
+    pub scale: f64,
 }
 
 impl Checkered {
-    pub fn texture(scale: f32, even: Vector3<f32>, odd: Vector3<f32>) -> Texture {
+    pub fn texture(scale: f64, even: Vector3<f64>, odd: Vector3<f64>) -> Texture {
         let scale = 1.0 / scale;
         Texture::Checkered(Checkered { scale, even, odd })
     }
 }
 
 impl Sample for Checkered {
-    fn sample(&self, _: f32, _: f32, p: Vector3<f32>) -> Vector3<f32> {
+    fn sample(&self, _: f64, _: f64, p: Vector3<f64>) -> Vector3<f64> {
         let x_int = (p.x * self.scale).floor() as i32;
         let y_int = (p.y * self.scale).floor() as i32;
         let z_int = (p.z * self.scale).floor() as i32;
@@ -85,31 +85,31 @@ impl Image {
 }
 
 impl Sample for Image {
-    fn sample(&self, u: f32, v: f32, _: Vector3<f32>) -> Vector3<f32> {
+    fn sample(&self, u: f64, v: f64, _: Vector3<f64>) -> Vector3<f64> {
         let interval = Interval::new(0.0, 1.0);
 
         let u = interval.clamp(u);
         let v = interval.clamp(v);
 
-        let i = (u * self.data.width() as f32) as u32;
-        let j = (v * self.data.height() as f32) as u32;
+        let i = (u * self.data.width() as f64) as u32;
+        let j = (v * self.data.height() as f64) as u32;
 
         let pixel = self.data.get_pixel(i, j);
         let color_scale = 1.0 / 256.0;
 
-        Vector3::new(pixel[0] as f32, pixel[1] as f32, pixel[2] as f32) * color_scale
+        Vector3::new(pixel[0] as f64, pixel[1] as f64, pixel[2] as f64) * color_scale
     }
 }
 
 #[derive(Debug, Clone)]
 pub struct Noise {
     pub perlin: Perlin,
-    pub scale: f32,
+    pub scale: f64,
     pub turbulance: u32,
 }
 
 impl Noise {
-    pub fn texture(scale: f32, turbulance: u32) -> Texture {
+    pub fn texture(scale: f64, turbulance: u32) -> Texture {
         let perlin = Perlin::default();
         Texture::Noise(Noise {
             perlin,
@@ -120,7 +120,7 @@ impl Noise {
 }
 
 impl Sample for Noise {
-    fn sample(&self, _: f32, _: f32, p: Vector3<f32>) -> Vector3<f32> {
-        Vector3::from_element(1.0f32) * (self.perlin.turb(p * self.scale, self.turbulance))
+    fn sample(&self, _: f64, _: f64, p: Vector3<f64>) -> Vector3<f64> {
+        Vector3::from_element(1.0f64) * (self.perlin.turb(p * self.scale, self.turbulance))
     }
 }
